Add New Features: New Family Names and First Names
-     Updates in Family and First Names for 17 Different Languages or Geographical (Grouped) Areas.

Disclaimer: Most of the names and surnames have been gathered using ChatGPT. Please consider that the author is not an expert in different languages, and some groups might seem inappropriate to native speakers. Therefore, the author cannot guarantee the absence of errors in this database. You are welcome to contact the author with any suggestions for modifications.