Add New Feature: Language Carrousel Selector Linked to other Object (Card)
The item selected in the carrousel is now linked to other components (in this case a Card). This way we have a visual selector and inside the card we can input more data.

Since the Carrousel is only able to return the active item (int), we are not able to access other data such as its "key", or "header". The way to link these objects (Carrousel and Card) is quite interesting: they are linked through a non-visible Div that returns a json (so we can use it as input to the Card update callback).