Add New Feature: Name's Ethnic-Origin Slidebar
Now you can choose the proportion of students with a name belonging to a ethnicity group in the Slidebar. The content the user chooses in now saved to the JSON requirements.

TODO: Read from the JSON to refresh and not overwrite values when the user selects anothe main language in the carousel.