Enhanced Features
In the "number of students" selector, if the user input a value outside of the range, the input box changes its appearence (making clear that the input is invalid).

The (gender) "bias selector plot" is now responsive to the number of students and the fe/male proportion the user inputs.