Fix Labelling Bug
When labelled as "other", the last word of any segment would not be included in the "words" attribute of the segment. This behaviour was even more critical in those cases where the segment was only composed by one word. The bug in now solved.

In addition, we include a bbox visualization module to make sure the bboxes given by the pipeline are correct.