Debug Data Gathering to Only Include Subsets in Desired Combination
Previously, the architecture allowed for moving data in and out of the train-val folder; however, every subset in app/data (including "train-val", "bruteforce" for training/validation, and "test" for testing) was loaded during the training session.

Now, only the "train-val" and "test" folders are loaded, while the "bruteforce" folder serves solely as a storage location accessed only during preprocessing steps as needed.